+++ lwb/metalang/spt/spt/src/main/java/mb/spt/expectation/CheckPatternExpectation.java
@@ -71,7 +71,7 @@ public class CheckPatternExpectation implements TestExpectation {
         }
 
         if(addMessages) {
-            SptMessageRemap.addMessagesRemapped(messagesBuilder, testCase.resource, file, result);
+            SptMessageRemap.addMessagesRemapped(messagesBuilder, testCase, file, sourceRegion, result);
         }
 
         return messagesBuilder.build(file);