+++ lwb/metalang/spt/spt/src/main/java/mb/spt/expectation/TransformToAtermExpectation.java
@@ -91,7 +91,7 @@ public class TransformToAtermExpectation implements TestExpectation {
         }
 
         if(feedback.hasErrorMessages()) {
-            SptMessageRemap.addMessagesRemapped(messagesBuilder, testCase.resource, file, feedback.getMessages());
+            SptMessageRemap.addMessagesRemapped(messagesBuilder, testCase, file, sourceRegion, feedback.getMessages());
         }
 
         return messagesBuilder.build(file);