+++ lwb/metalang/spt/spt/src/main/java/mb/spt/expectation/TransformToFragmentExpectation.java
@@ -80,7 +80,7 @@ public class TransformToFragmentExpectation implements TestExpectation {
         }
 
         if(feedback.hasErrorMessages()) {
-            SptMessageRemap.addMessagesRemapped(messagesBuilder, testCase.resource, file, feedback.getMessages());
+            SptMessageRemap.addMessagesRemapped(messagesBuilder, testCase, file, sourceRegion, feedback.getMessages());
         }
 
         return messagesBuilder.build(file);