+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project are documented in this file, based on [Keep 
 ### Fixed
 - More wrong main files being used when only the main source directory was set for a meta-language in `spoofaxc.cfg`.
 - Extract more messages in SPT `transform` expectations.
+- Out-of-bounds messages in SPT. Messages outside the bounds of the test fragment are now moved to the expectation.
 
 
 ## [0.16.11] - 2021-11-19